before:
```
hyperfine --warmup 2 --export-csv result-queries.csv \
	'make run query=get-periodic-total-failure-rate' 
Benchmark 1: make run query=get-periodic-total-failure-rate
  Time (mean ± σ):     393.2 ms ±  11.5 ms    [User: 333.8 ms, System: 60.5 ms]
  Range (min … max):   384.0 ms … 424.6 ms    10 runs
```

after
```
hyperfine --warmup 2 --export-csv result-queries.csv \
	'make run query=get-periodic-total-failure-rate' 
Benchmark 1: make run query=get-periodic-total-failure-rate
  Time (mean ± σ):     163.1 ms ±   7.1 ms    [User: 129.6 ms, System: 35.1 ms]
  Range (min … max):   154.9 ms … 173.2 ms    17 runs
```